Fix cwtools install on new laptop.
After much experimentation, it looks like the File.rename() was
failing due to the unzip of Cygwin.zip keeping hold of the
directory. After adding a delay before the rename, it now
seems to work. It's not nice to add a delay, but I don't have
time at the moment to add and rename retry logic.